refactor(ios): extract SheetChip and SheetDismissButton (#256)
Both were always shared — SheetChip by FilterSheet and FacetChipCloud,
SheetDismissButton by FilterSheet — and only lived in DateFilterSheet.swift
because that is where the first one was written. Moving them out is a
prerequisite for deleting that file.

Pure move. No type, doc comment, or behaviour changed.
